refactor: extract NodeTextArea from EditableNode
mainly because there's even more textarea-specific logic being
added soon, but there's already a lot of it, so seems good to do anyway.
